backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.30.6
websockets==13.1
redis[hiredis]==5.0.8
pydantic[email]==2.8.2
pydantic-settings==2.4.0
PyJWT==2.12.0
passlib==1.7.4
httpx==0.27.2
slowapi==0.1.9
python-dotenv==1.0.1
email-validator==2.2.0
bleach==6.1.0
python-multipart==0.0.22
starlette==0.49.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
sentry-sdk[fastapi]==2.19.2

# -- Testing ------------------------------------------------------------------
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-httpx==0.34.0
respx==0.22.0
